Split NNPages into focused modules
